+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - **Crypt:** Added support for `node:crypto` API.
 - **KDF:** Added modules `@astrobase/sdk/kdf/node` & `@astrobase/sdk/kdf/web-crypto`.
 - **KDF:** Added support for using a raw input material for KDF.
+- **SQLite:** Added support for passing a `Database` instance instead of a config.
 
 ### Developer